Reorganize api
