+++ backend/app/storage.py
@@ -61,11 +61,11 @@ def list_assistants(user_id: str) -> List[Assistant]:
     return [load(assistant_hash_keys, values) for values in assistants]
 
 
-def get_assistant(user_id: str, assistant_id: str) -> Assistant:
+def get_assistant(user_id: str, assistant_id: str) -> Assistant | None:
     """Get an assistant by ID."""
     client = _get_redis_client()
     values = client.hmget(assistant_key(user_id, assistant_id), *assistant_hash_keys)
-    return load(assistant_hash_keys, values)
+    return load(assistant_hash_keys, values) if any(values) else None
 
 
 def list_public_assistants(
