+++ frontend/vite.config.ts
@@ -6,12 +6,11 @@ export default defineConfig({
   plugins: [react()],
   server: {
     proxy: {
-      "^/(config_schema|input_schema|stream|assistants|threads|ingest|feedback|runs)":
-        {
-          target: "http://127.0.0.1:8100",
-          changeOrigin: true,
-          rewrite: (path) => path.replace("/____LANGSERVE_BASE_URL", ""),
-        },
+      "^/(assistants|threads|ingest|runs)": {
+        target: "http://127.0.0.1:8100",
+        changeOrigin: true,
+        rewrite: (path) => path.replace("/____LANGSERVE_BASE_URL", ""),
+      },
     },
   },
 });
